Move counting strategies to sub-package
