+++ contentgrid-spring-data-pagination/src/main/java/com/contentgrid/spring/data/pagination/jpa/ContentGridSpringDataPaginationJpaConfiguration.java
@@ -1,5 +1,9 @@
 package com.contentgrid.spring.data.pagination.jpa;
 
+import com.contentgrid.spring.data.pagination.jpa.strategy.AggregateItemCountStrategy;
+import com.contentgrid.spring.data.pagination.jpa.strategy.ExplainEstimateItemCountStrategy;
+import com.contentgrid.spring.data.pagination.jpa.strategy.JpaQuerydslItemCountStrategy;
+import com.contentgrid.spring.data.pagination.jpa.strategy.TimedDirectCountItemCountStrategy;
 import java.util.List;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
